+++ String Generation/RegressionStringGenerator/QueryInfo.cs
@@ -1,6 +1,6 @@
 ﻿namespace citynames;
-public class QueryInfo(IReadOnlyDictionary<string, float> biomeWeights)
+public class QueryInfo(IReadOnlyDictionary<string, double> biomeWeights)
 {
-    public IReadOnlyDictionary<string, float> BiomeWeights = biomeWeights;
+    public IReadOnlyDictionary<string, double> BiomeWeights = biomeWeights;
     public QueryInfo(string biome) : this(biome.ToWeightVector()) { }
 }
